+++ src/Material/PsychMat.hpp
@@ -10,5 +10,5 @@ class PsychMat : public RayTracer::I_Material {
     ~PsychMat() override = default;
 
     void Init() override;
-    sf::Color getColorAt(int x, int y) override;
+    sf::Color getColorAt(float x, float y) override;
 };
